+++ jest.config.js
@@ -1,0 +1,13 @@
+const nextJest = require('next/jest')
+
+const createJestConfig = nextJest()
+
+// Any custom config you want to pass to Jest
+const customJestConfig = {
+  testMatch: ['**/*.test.js', '**/*.test.ts', '**/*.test.tsx'],
+  verbose: true,
+  transformIgnorePatterns: ['/next[/\\\\]dist/', '/\\.next/'],
+}
+
+// createJestConfig is exported in this way to ensure that next/jest can load the Next.js config which is async
+module.exports = createJestConfig(customJestConfig)
