+++ package.json
@@ -20,8 +20,7 @@
     "lint": "tsdx lint",
     "prepare": "tsc"
   },
-  "dependencies": {
-  },
+  "dependencies": {},
   "peerDependencies": {
     "next": ">=12"
   },
@@ -45,14 +44,17 @@
     "url": "git+https://github.com/labd/nextjs-basic-auth-middleware"
   },
   "devDependencies": {
+    "@edge-runtime/jest-environment": "^1.1.0-beta.26",
     "@types/jest": "26.0.15",
     "@types/node": "14.14.8",
+    "edge-mock": "^0.0.15",
     "husky": "4.3.0",
+    "jest": "^28.1.3",
+    "next": "12.1.6",
+    "node-fetch": "3.2.4",
     "node-mocks-http": "1.9.0",
     "tsdx": "0.14.1",
     "tslib": "2.4.0",
-    "typescript": "4.7.3",
-    "next": "12.1.6",
-    "node-fetch": "3.2.4"
+    "typescript": "4.7.3"
   }
 }
