+++ src/index.ts
@@ -1,141 +1,3 @@
-import { IncomingMessage, ServerResponse } from 'http'
-import { NextRequest, NextResponse } from 'next/server'
-
-import {
-  compareCredentials,
-  parseCredentials,
-  AuthCredentials,
-} from './credentials.js'
-import { pathInRequest } from './path.js'
-import { basicAuthentication } from './lib/auth.js'
-
-export type MiddlewareOptions = {
-  realm?: string
-  users?: AuthCredentials
-  includePaths?: string[]
-  excludePaths?: string[]
-}
-
-/**
- * Middleware that sends a basic auth challenge to the user when enabled
- * @param req Http server incoming message
- * @param res Server response
- */
-export const pageMiddleware = (
-  req: IncomingMessage,
-  res: ServerResponse,
-  {
-    realm = 'protected',
-    users = [],
-    includePaths = ['/'],
-    excludePaths = [],
-  }: MiddlewareOptions = {}
-) => {
-  if (!req?.url) {
-    // Current request object has no url defined so cannot set up basic authentication
-    // This means page does not have a getServerSideProps or getInitialProps function
-    return
-  }
-  // Check if credentials are set up
-  const environmentCredentials = process.env.BASIC_AUTH_CREDENTIALS || ''
-  if (environmentCredentials.length === 0 && users.length === 0) {
-    // No credentials set up, continue rendering the page as normal
-    return
-  }
-
-  // Retrieve paths from environment credentials or use arguments
-  const includeAuth = process.env.BASIC_AUTH_PATHS
-    ? process.env.BASIC_AUTH_PATHS.split(';')
-    : includePaths
-  const excludeAuth = process.env.BASIC_AUTH_EXCLUDE_PATHS
-    ? process.env.BASIC_AUTH_EXCLUDE_PATHS.split(';')
-    : excludePaths
-
-  // Check whether the path of the request should even be checked
-  if (pathInRequest(excludeAuth, req) || !pathInRequest(includeAuth, req)) {
-    // Current path not part of the checked settings
-    return
-  }
-
-  const credentialsObject: AuthCredentials =
-    environmentCredentials.length > 0
-      ? parseCredentials(environmentCredentials)
-      : users
-
-  const currentUser = basicAuthentication(req.headers.authorization)
-  if (!currentUser || !compareCredentials(currentUser, credentialsObject)) {
-    res.statusCode = 401
-    res.setHeader('WWW-Authenticate', `Basic realm="${realm}"`)
-    res.end('401 Access Denied')
-  }
-}
-
-/**
- * Creates a default Next middleware function that returns `NextResponse.next()` if the basic auth passes
- * @param req Next middleware request
- * @param options Options object based on MiddlewareOptions
- * @returns Either a 401 error or goes to the next page
- */
-export const createNextMiddleware = ({
-  realm = 'protected',
-  users = [],
-  includePaths = ['/'],
-  excludePaths = [],
-}: MiddlewareOptions = {}) => (req: NextRequest) => {
-  nextBasicAuthMiddleware({ realm, users, includePaths, excludePaths }, req)
-
-  return NextResponse.next()
-}
-
-export const nextBasicAuthMiddleware = (
-  {
-    realm = 'protected',
-    users = [],
-    includePaths = ['/'],
-    excludePaths = [],
-  }: MiddlewareOptions = {},
-  req: NextRequest
-) => {
-  // Check if credentials are set up
-  const environmentCredentials = process.env.BASIC_AUTH_CREDENTIALS || ''
-  if (environmentCredentials.length === 0 && users.length === 0) {
-    // No credentials set up, continue rendering the page as normal
-    return
-  }
-
-  // Retrieve paths from environment credentials or use arguments
-  const includeAuth = process.env.BASIC_AUTH_PATHS
-    ? process.env.BASIC_AUTH_PATHS.split(';')
-    : includePaths
-  const excludeAuth = process.env.BASIC_AUTH_EXCLUDE_PATHS
-    ? process.env.BASIC_AUTH_EXCLUDE_PATHS.split(';')
-    : excludePaths
-
-  // Check whether the path of the request should even be checked
-  if (pathInRequest(excludeAuth, req) || !pathInRequest(includeAuth, req)) {
-    // Current path not part of the checked settings
-    return
-  }
-
-  const credentialsObject: AuthCredentials =
-    environmentCredentials.length > 0
-      ? parseCredentials(environmentCredentials)
-      : users
-
-  const authHeader = req.headers.get('authorization')
-
-  if (authHeader) {
-    const currentUser = basicAuthentication(authHeader)
-
-    if (currentUser && compareCredentials(currentUser, credentialsObject)) {
-      return
-    }
-  }
-
-  return new Response('401 Access Denied', {
-    status: 401,
-    headers: {
-      'WWW-Authenticate': `Basic realm="${realm}`,
-    },
-  })
-}
+export { createNextMiddleware, nextBasicAuthMiddleware } from './nextMiddleware'
+export { serverMiddleware } from './serverMiddleware'
+export type { MiddlewareOptions } from './types'
