+++ src/types.d.ts
@@ -1,0 +1,8 @@
+import { AuthCredentials } from 'lib/credentials'
+
+export type MiddlewareOptions = {
+  realm?: string
+  users?: AuthCredentials
+  includePaths?: string[]
+  excludePaths?: string[]
+}
